requirements.txt
requests
pymupdf
scikit-learn
reportlab
playwright
numpy
python-dotenv
langgraph
langchain
langchain-ollama
